+++ src/main/java/com/github/alexthe666/iceandfire/entity/EntityDragonSkull.java
@@ -53,14 +53,14 @@ public class EntityDragonSkull extends EntityAnimal implements IBlacklistedFromS
 	@Override
 	protected void entityInit() {
 		super.entityInit();
-		this.getDataManager().register(DRAGON_TYPE, 0);
-		this.getDataManager().register(DRAGON_AGE, 0);
-		this.getDataManager().register(DRAGON_STAGE, 0);
-		this.getDataManager().register(DRAGON_DIRECTION, 0.0f);
+		this.getDataManager().register(DRAGON_TYPE, Integer.valueOf(0));
+		this.getDataManager().register(DRAGON_AGE, Integer.valueOf(0));
+		this.getDataManager().register(DRAGON_STAGE, Integer.valueOf(0));
+		this.getDataManager().register(DRAGON_DIRECTION, Float.valueOf(0F));
 	}
 
 	public float getYaw() {
-		return this.getDataManager().get(DRAGON_DIRECTION);
+		return this.getDataManager().get(DRAGON_DIRECTION).floatValue();
 	}
 
 	public void setYaw(float var1) {
@@ -68,7 +68,7 @@ public class EntityDragonSkull extends EntityAnimal implements IBlacklistedFromS
 	}
 
 	public int getType() {
-		return this.getDataManager().get(DRAGON_TYPE);
+		return this.getDataManager().get(DRAGON_TYPE).intValue();
 	}
 
 	public void setType(int var1) {
@@ -76,7 +76,7 @@ public class EntityDragonSkull extends EntityAnimal implements IBlacklistedFromS
 	}
 
 	public int getStage() {
-		return this.getDataManager().get(DRAGON_STAGE);
+		return this.getDataManager().get(DRAGON_STAGE).intValue();
 	}
 
 	public void setStage(int var1) {
@@ -84,7 +84,7 @@ public class EntityDragonSkull extends EntityAnimal implements IBlacklistedFromS
 	}
 
 	public int getDragonAge() {
-		return this.getDataManager().get(DRAGON_AGE);
+		return this.getDataManager().get(DRAGON_AGE).intValue();
 	}
 
 	public void setDragonAge(int var1) {
