+++ src/main/java/com/github/alexthe666/iceandfire/entity/EntityTroll.java
@@ -101,12 +101,12 @@ public class EntityTroll extends EntityMob implements IAnimatedEntity {
     @Override
     protected void entityInit() {
         super.entityInit();
-        this.dataManager.register(VARIANT, 0);
-        this.dataManager.register(WEAPON, 0);
+        this.dataManager.register(VARIANT, Integer.valueOf(0));
+        this.dataManager.register(WEAPON, Integer.valueOf(0));
     }
 
     private int getVariant() {
-        return this.dataManager.get(VARIANT);
+        return this.dataManager.get(VARIANT).intValue();
     }
 
     private void setVariant(int variant) {
@@ -122,7 +122,7 @@ public class EntityTroll extends EntityMob implements IAnimatedEntity {
     }
 
     private int getWeapon() {
-        return this.dataManager.get(WEAPON);
+        return this.dataManager.get(WEAPON).intValue();
     }
 
     private void setWeapon(int variant) {
